lesson 7 compleated
-player controller improved
-smart enemy added
-all int variable chanced float except rectangles

+++ src/com/tutorial/main/BasicEnemy.java
@@ -21,7 +21,7 @@ public class BasicEnemy extends GameObject {
 		}
 	
 	public Rectangle getBounds(){
-		return new Rectangle(x, y, 16, 16);
+		return new Rectangle((int)x, (int)y, 16, 16);
 	}
 
 	public void tick() {
@@ -39,7 +39,7 @@ public class BasicEnemy extends GameObject {
 
 	public void render(Graphics g) {
 		g.setColor(Color.RED);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int)x,(int) y, 16, 16);
