+++ src/com/tutorial/main/Game.java
@@ -101,12 +101,13 @@ public class Game extends Canvas implements Runnable{
 		bs.show();
 	}
 	
-	public static int clamp(int var ,int min ,int  max){
+	public static float clamp(float var ,float min ,float  max){
 		if (var >= max)
 			return var = max ;
 		else if (var <= min )
 			return var = min ;
-		else  return var ;
+		else  
+			return var ;
 	}
 	
 	public static void main (String []args){
